api/requirements.txt
aiodataloader==0.1.2
cloudant==2.10.0
Flask==1.0.2
flask-graphql==2.0.0
google-cloud-logging==1.6.0
graphene==2.1.3
gunicorn==19.7.1
requests==2.18.4
web3==4.7.2
